1. Added a algorithm for Money Change which solves the problem: The goal in this problem is to find the minimum number of coins needed to change the input value (an integer) into coins with denominations 1, 5, and 10.
2. Also added empty project for GroupChildren, which solves the problem: You've invited a lot of children to a celebration party, and
you want to entertain them and also teach them something in the process.
You are going to hire a few teachers and
divide the children into groups and assign a teacher to each of the groups
this teacher will work with this group through the whole party. But you know that for a teacher to work with a group of children efficiently
children of that group should be of relatively the same age.
More specifically age of any two children in the same group
should differ by at most one year.  Also, you want to minimize the number of groups.
Because you want to hire fewer teachers, and
spend the money on presents and other kinds of entertainment for the children.
So, you need to divide children into the minimum possible number of groups.
Such that the age of any two children in any group differs by at most one year.